+++ src/backend/resolvers/mutations/client/index.js
@@ -1,5 +1,9 @@
 const createClient = require('./create')
+const deleteClient = require('./delete')
+const updateClient = require('./update')
 
 module.exports = {
   createClient,
+  deleteClient,
+  updateClient,
 }