+++ src/backend/typeDefs/mutations/client/delete.js
@@ -1,0 +1,15 @@
+const { gql } = require('apollo-server-express')
+
+const deleteClientTypeDefs = gql`
+  input DeleteClientInput {
+    _id: ID!
+  }
+
+  type DeleteClientPayload {
+    _id: ID!
+    name: String
+    description: String
+  }
+`
+
+module.exports = deleteClientTypeDefs