+++ src/backend/typeDefs/mutations/client/index.js
@@ -1,5 +1,9 @@
 const createClientTypeDefs = require('./create')
+const deleteClientTypeDefs = require('./delete')
+const updateClientTypeDefs = require('./update')
 
 module.exports = [
   createClientTypeDefs,
+  deleteClientTypeDefs,
+  updateClientTypeDefs,
 ]