+++ src/backend/typeDefs/mutations/client/update.js
@@ -1,0 +1,16 @@
+const { gql } = require('apollo-server-express')
+
+const updateClientTypeDefs = gql`
+  input UpdateClientInput {
+    _id: ID!
+    description: String!
+  }
+
+  type UpdateClientPayload {
+    _id: ID!
+    name: String
+    description: String
+  }
+`
+
+module.exports = updateClientTypeDefs