+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Phoenix`: `upsertUsersPermissions` and `upsertUsersSitemaps` utils for persisting user access directly to `dev`
 - `Phoenix`: `getUser` method was added to Auth0 Management Client and used in `auth0/index.js` instead of the method in Auth0 Extension Client to pave way for auth0 ext client's removal
+- `Phoenix`: Add Client update functionality (and build out delete functionality but leave it hidden)
 
 ### Changed
 - `Delphi`: refactored backend to be more flexible with less hard-coding for Pathways subscription.